requirements.txt
django_extensions
channels